+++ programs/drift/src/controller/repeg.rs
@@ -4,7 +4,6 @@ use anchor_lang::prelude::AccountInfo;
 use anchor_lang::prelude::*;
 use solana_program::msg;
 
-use crate::controller::amm::calculate_perp_market_amm_summary_stats;
 use crate::controller::amm::update_spreads;
 use crate::controller::spot_balance::update_spot_balances;
 use crate::error::ErrorCode;
@@ -423,9 +422,6 @@ pub fn settle_expired_market(
         "target_expiry_price <= 0 {}",
         target_expiry_price
     )?;
-
-    let total_excess_balance2: i128 =
-        calculate_perp_market_amm_summary_stats(market, spot_market, target_expiry_price)?;
 
     let pnl_pool_token_amount = get_token_amount(
         market.pnl_pool.scaled_balance,