+++ programs/drift/src/instructions/admin.rs
@@ -41,6 +41,7 @@ use crate::state::fulfillment_params::phoenix::PhoenixMarketContext;
 use crate::state::fulfillment_params::phoenix::PhoenixV1FulfillmentConfig;
 use crate::state::fulfillment_params::serum::SerumContext;
 use crate::state::fulfillment_params::serum::SerumV3FulfillmentConfig;
+use crate::state::high_leverage_mode_config::HighLeverageModeConfig;
 use crate::state::insurance_fund_stake::ProtocolIfSharesTransferConfig;
 use crate::state::oracle::get_sb_on_demand_price;
 use crate::state::oracle::{
@@ -803,6 +804,8 @@ pub fn handle_initialize_perp_market(
     validate_margin(
         margin_ratio_initial,
         margin_ratio_maintenance,
+        0,
+        0,
         liquidator_fee,
         max_spread,
     )?;
@@ -855,7 +858,10 @@ pub fn handle_initialize_perp_market(
         fuel_boost_position: 0,
         fuel_boost_taker: 0,
         fuel_boost_maker: 0,
-        padding: [0; 43],
+        padding1: 0,
+        high_leverage_margin_ratio_initial: 0,
+        high_leverage_margin_ratio_maintenance: 0,
+        padding: [0; 38],
         amm: AMM {
             oracle: *ctx.accounts.oracle.key,
             oracle_source,
@@ -2227,6 +2233,8 @@ pub fn handle_update_perp_market_margin_ratio(
     validate_margin(
         margin_ratio_initial,
         margin_ratio_maintenance,
+        perp_market.high_leverage_margin_ratio_initial.cast()?,
+        perp_market.high_leverage_margin_ratio_maintenance.cast()?,
         perp_market.liquidator_fee,
         perp_market.amm.max_spread,
     )?;
@@ -2245,6 +2253,48 @@ pub fn handle_update_perp_market_margin_ratio(
 
     perp_market.margin_ratio_initial = margin_ratio_initial;
     perp_market.margin_ratio_maintenance = margin_ratio_maintenance;
+    Ok(())
+}
+
+#[access_control(
+    perp_market_valid(&ctx.accounts.perp_market)
+)]
+pub fn handle_update_perp_market_high_leverage_margin_ratio(
+    ctx: Context<AdminUpdatePerpMarket>,
+    margin_ratio_initial: u16,
+    margin_ratio_maintenance: u16,
+) -> Result<()> {
+    let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+
+    msg!(
+        "updating perp market {} margin ratio",
+        perp_market.market_index
+    );
+
+    validate_margin(
+        perp_market.margin_ratio_initial,
+        perp_market.margin_ratio_maintenance,
+        margin_ratio_initial.cast()?,
+        margin_ratio_maintenance.cast()?,
+        perp_market.liquidator_fee,
+        perp_market.amm.max_spread,
+    )?;
+
+    msg!(
+        "perp_market.high_leverage_margin_ratio_initial: {:?} -> {:?}",
+        perp_market.high_leverage_margin_ratio_initial,
+        margin_ratio_initial
+    );
+
+    msg!(
+        "perp_market.high_leverage_margin_ratio_maintenance: {:?} -> {:?}",
+        perp_market.high_leverage_margin_ratio_maintenance,
+        margin_ratio_maintenance
+    );
+
+    perp_market.high_leverage_margin_ratio_initial = margin_ratio_initial;
+    perp_market.high_leverage_margin_ratio_maintenance = margin_ratio_maintenance;
+
     Ok(())
 }
 
@@ -2400,6 +2450,8 @@ pub fn handle_update_perp_liquidation_fee(
     validate_margin(
         perp_market.margin_ratio_initial,
         perp_market.margin_ratio_maintenance,
+        perp_market.high_leverage_margin_ratio_initial.cast()?,
+        perp_market.high_leverage_margin_ratio_maintenance.cast()?,
         liquidator_fee,
         perp_market.amm.max_spread,
     )?;
@@ -4092,6 +4144,35 @@ pub fn handle_settle_expired_market<'c: 'info, 'info>(
     Ok(())
 }
 
+pub fn handle_initialize_high_leverage_mode_config(
+    ctx: Context<InitializeHighLeverageModeConfig>,
+    max_users: u32,
+) -> Result<()> {
+    let mut config = ctx.accounts.high_leverage_mode_config.load_init()?;
+
+    config.max_users = max_users;
+
+    config.validate()?;
+
+    Ok(())
+}
+
+pub fn handle_update_high_leverage_mode_config(
+    ctx: Context<UpdateHighLeverageModeConfig>,
+    max_users: u32,
+    reduce_only: bool,
+) -> Result<()> {
+    let mut config = load_mut!(ctx.accounts.high_leverage_mode_config)?;
+
+    config.max_users = max_users;
+
+    config.reduce_only = reduce_only as u8;
+
+    config.validate()?;
+
+    Ok(())
+}
+
 #[derive(Accounts)]
 pub struct Initialize<'info> {
     #[account(mut)]
@@ -4701,6 +4782,42 @@ pub struct InitPythPullPriceFeed<'info> {
     #[account(mut, seeds = [PTYH_PRICE_FEED_SEED_PREFIX, &feed_id], bump)]
     pub price_feed: AccountInfo<'info>,
     pub system_program: Program<'info, System>,
+    #[account(
+        has_one = admin
+    )]
+    pub state: Box<Account<'info, State>>,
+}
+
+#[derive(Accounts)]
+pub struct InitializeHighLeverageModeConfig<'info> {
+    #[account(mut)]
+    pub admin: Signer<'info>,
+    #[account(
+        init,
+        seeds = [b"high_leverage_mode_config".as_ref()],
+        space = HighLeverageModeConfig::SIZE,
+        bump,
+        payer = admin
+    )]
+    pub high_leverage_mode_config: AccountLoader<'info, HighLeverageModeConfig>,
+    #[account(
+        has_one = admin
+    )]
+    pub state: Box<Account<'info, State>>,
+    pub rent: Sysvar<'info, Rent>,
+    pub system_program: Program<'info, System>,
+}
+
+#[derive(Accounts)]
+pub struct UpdateHighLeverageModeConfig<'info> {
+    #[account(mut)]
+    pub admin: Signer<'info>,
+    #[account(
+        mut,
+        seeds = [b"high_leverage_mode_config".as_ref()],
+        bump,
+    )]
+    pub high_leverage_mode_config: AccountLoader<'info, HighLeverageModeConfig>,
     #[account(
         has_one = admin
     )]