+++ sdk/src/driftClientConfig.ts
@@ -18,6 +18,7 @@ export type DriftClientConfig = {
 	wallet: IWallet;
 	env?: DriftEnv;
 	programID?: PublicKey;
+	swiftID?: PublicKey;
 	accountSubscription?: DriftClientSubscriptionConfig;
 	opts?: ConfirmOptions;
 	txSender?: TxSender;