+++ sdk/src/orderSubscriber/OrderSubscriber.ts
@@ -222,11 +222,8 @@ export class OrderSubscriber {
 					dataType
 				);
 			}
-			if (userAccount.hasOpenOrder) {
-				this.usersAccounts.set(key, { slot, userAccount });
-			} else {
-				this.usersAccounts.delete(key);
-			}
+
+			this.usersAccounts.set(key, { slot, userAccount });
 		}
 	}